Update dice roll visual feedback on web page
Added more meaningful visual iconography to each dice roll result type in the Roller.razor page. The changes allow users to see at a glance whether a roll was a critical success, critical fail, or dropped roll due to the addition of specific icons rather than just the text result. This should enhance the user experience, particularly for users with visual impairments relying on screen readers, as an 'alt' tag detailing the meaning of the icon has also been incorporated.